requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bottleneck=1.3.5=py38h7deecbd_0
ca-certificates=2023.12.12=h06a4308_0
cudatoolkit=11.0.221=h6bb024c_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.3.2=pypi_0
jpeg=9b=h024ee3a_2
lcms2=2.12=h3be6417_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=h85742a9_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.4=h6a678d5_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
ncurses=6.4=h6a678d5_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numexpr=2.8.4=py38he184ba9_0
numpy=1.24.4=pypi_0
numpy-base=1.24.3=py38h31eccc5_0
openssl=1.1.1w=h7f8727e_0
packaging=23.1=py38h06a4308_0
pandas=2.0.3=py38h1128e8f_0
pillow=9.3.0=py38hace64e9_1
pip=23.3.1=py38h06a4308_0
python=3.8.0=h0371630_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pytz=2023.3.post1=py38h06a4308_0
readline=7.0=h7b6447c_5
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu110
typing_extensions=4.7.1=py38h06a4308_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.5=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
